Iteracio 5 projectes en menu (#111)
* Afegir projectes to menu

* Afegir id al esborrar

* Afegir projectes al menu nomes autentificats

* Typo

* Adaptar links a la nova url de edt_project amb id

* Actualitzar edit_project amb id

* Ajustes en invitation

* Canvi text

* Text i eliminar OPEN del Home

---------

+++ src/apps/coopolis/views/ProjectFormView.py
@@ -1,4 +1,3 @@
-from apps.coopolis.models.invitation import Invitation
 from conf import settings
 from conf.custom_mail_manager import MyMailTemplate
 from constance import config
@@ -9,10 +8,14 @@ from django.shortcuts import redirect, render
 from django.contrib.messages.views import SuccessMessageMixin
 from django.http import HttpResponseRedirect
 from django.views import generic
+from django.shortcuts import get_object_or_404
+from django.core.validators import EmailValidator
+from django.core.exceptions import ValidationError
 
 from apps.cc_courses.choices import ProjectStageStatesChoices
 from apps.coopolis.forms import ProjectForm
 from apps.coopolis.models import Project, ProjectStage, User
+from apps.coopolis.models.invitation import Invitation
 from apps.coopolis.views import LoginSignupContainerView
 
 
@@ -23,15 +26,12 @@ class ProjectFormView(SuccessMessageMixin, generic.UpdateView):
     success_message = "Dades del projecte actualitzades correctament."
 
     def get_success_url(self):
-        return urls.reverse("edit_project")
+        return urls.reverse("edit_project", kwargs={"pk": self.object.pk})
 
     def get_object(self, queryset=None):
-        return self.request.user.project
-
-    def get(self, request):
-        if self.request.user.project is None:
-            return HttpResponseRedirect(urls.reverse("new_project"))
-        return super().get(self, request)
+        return get_object_or_404(
+            Project, pk=self.kwargs["pk"], partners__id=self.request.user.id
+        )
 
     def post(self, request, *args, **kwargs):
         if "delete" in request.POST:
@@ -45,12 +45,15 @@ class ProjectFormView(SuccessMessageMixin, generic.UpdateView):
                 request,
                 "Dades de l'acompanyament esborrades correctament.",
             )
-            return HttpResponseRedirect(urls.reverse("edit_project"))
+            return HttpResponseRedirect(
+                urls.reverse("edit_project", kwargs={"pk": project.pk})
+            )
         return super().post(request, *args, **kwargs)
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        project = self.request.user.project
+        object_id = self.kwargs.get("pk")
+        project = Project.objects.get(pk=object_id)
         pending_project_stages = ProjectStage.objects.filter(
             project=project, stage_state=ProjectStageStatesChoices.PENDING
         )
@@ -65,42 +68,52 @@ class ProjectFormView(SuccessMessageMixin, generic.UpdateView):
         users = set()
         if invited_users:
             for user in invited_users:
-                users.add(User.objects.filter(id=user['user']).first())
+                users.add(User.objects.filter(id=user["user"]).first())
         context["invited_partners"] = users
         return context
 
 
 @login_required
 def project_partner_manage(request):
+    project_pk = None
     if request.method == "POST":
         if "add_partner" in request.POST:
-            if not request.POST.get("email"):
+            project_pk = request.POST.get("add_partner")
+            project = Project.objects.get(pk=project_pk)
+            validator = EmailValidator()
+            try:
+                validator(request.POST.get("email"))
+            except ValidationError:
                 messages.error(
                     request,
                     "Ha d'indicar un correu electrònic vàlid.",
                 )
-                return redirect("edit_project")
+                return redirect("edit_project", pk=project_pk)
             try:
                 user = User.objects.get(email=request.POST.get("email"))
             except User.DoesNotExist:
                 messages.error(
                     request,
-                    "L'usuari no existeix.", )
-                return redirect("edit_project")
-            user_is_invited = Invitation.objects.filter(user=user).values("user__id").first()
+                    "L'usuari no existeix.",
+                )
+                return redirect("edit_project", pk=project_pk)
+            user_is_invited = (
+                Invitation.objects.filter(user=user, project=project)
+                .values("user__id")
+                .first()
+            )
             if user_is_invited:
                 messages.error(
                     request,
                     f"L'usuari {user.full_name} ja està invitat.",
                 )
-                return redirect("edit_project")
-            if user.project:
+                return redirect("edit_project", pk=project_pk)
+            if project.partners.filter(pk=user.pk).exists():
                 messages.error(
                     request,
                     f"L'usuari {user.full_name} ja forma part d'aquest projecte.",
                 )
-                return redirect("edit_project")
-            project = Project.objects.get(id=request.POST.get("add_partner"))
+                return redirect("edit_project", pk=project_pk)
             invitation = Invitation.objects.create(user=user, project=project)
             mail = MyMailTemplate("EMAIL_PROJECT_INVITATION")
             mail.to = user.email
@@ -110,7 +123,7 @@ def project_partner_manage(request):
                 "persona_email": user.email,
                 "project": project.name,
                 "absolute_url": settings.ABSOLUTE_URL,
-                "invitation_url": f"{settings.ABSOLUTE_URL}/project/invitation/{str(invitation.uuid)}"
+                "invitation_url": f"{settings.ABSOLUTE_URL}/project/invitation/{str(invitation.uuid)}",
             }
             mail.send(now=True)
             messages.success(
@@ -119,6 +132,15 @@ def project_partner_manage(request):
                 " Se li ha enviat un correu per a la seva acceptació.",
             )
         if "delete_partner" in request.POST:
+            project_pk = request.POST.get("delete_partner")
+            try:
+                project = Project.objects.get(pk=project_pk)
+            except Project.DoesNotExist:
+                messages.error(
+                    request,
+                    "Aquest projecte no existeix.",
+                )
+                return redirect("home")
             try:
                 user = User.objects.get(id=request.POST.get("partner_id"))
             except User.DoesNotExist:
@@ -126,21 +148,13 @@ def project_partner_manage(request):
                     request,
                     "Aquest usuari no existeix.",
                 )
-                return redirect("edit_project")
+                return redirect("edit_project", pk=project.pk)
             if user == request.user:
                 messages.error(
                     request,
                     "No et pots eliminar a tu mateix/a del projecte.",
                 )
-                return redirect("edit_project")
-            try:
-                project = Project.objects.get(pk=request.POST.get("delete_partner"))
-            except Project.DoesNotExist:
-                messages.error(
-                    request,
-                    "Aquest projecte no existeix.",
-                )
-                return redirect("edit_project")
+                return redirect("edit_project", pk=project.pk)
             project.partners.remove(user)
 
             mail = MyMailTemplate("EMAIL_PARTNER_ELIMINATION")
@@ -154,31 +168,34 @@ def project_partner_manage(request):
             mail.send(now=True)
             messages.success(
                 request,
-                f"{user.full_name} ha estat eliminada amb èxit d'aquest projecte.", )
-
-    if "delete_invitation" in request.POST:
-        try:
-            user = User.objects.get(id=request.POST.get("invited_id"))
-        except User.DoesNotExist:
-            messages.error(
-                request,
-                "Aquest usuari no existeix.",
+                f"{user.full_name} ha estat eliminada amb èxit d'aquest projecte.",
             )
-            return redirect("edit_project")
-        try:
-            invitation = Invitation.objects.get(user=user, project=request.POST.get("delete_invitation"))
-        except Invitation.DoesNotExist:
-            messages.error(
+        if "delete_invitation" in request.POST:
+            project_pk = request.POST.get("delete_invitation")
+            try:
+                user = User.objects.get(id=request.POST.get("invited_id"))
+            except User.DoesNotExist:
+                messages.error(
+                    request,
+                    "Aquest usuari no existeix.",
+                )
+                return redirect("edit_project", pk=project_pk)
+            try:
+                invitation = Invitation.objects.get(
+                    user=user, project=request.POST.get("delete_invitation")
+                )
+            except Invitation.DoesNotExist:
+                messages.error(
+                    request,
+                    "Aquesta invitació no existeix.",
+                )
+                return redirect("edit_project", pk=project_pk)
+            invitation.delete()
+            messages.success(
                 request,
-                "Aquesta invitació no existeix.",
+                f"{user.full_name} ha estat eliminada amb èxit d'aquest projecte.",
             )
-            return redirect("edit_project")
-        invitation.delete()
-        messages.success(
-            request,
-            f"{user.full_name} ha estat eliminada amb èxit d'aquest projecte.",
-        )
-    return redirect("edit_project")
+    return redirect("edit_project", pk=project_pk)
 
 
 @login_required
@@ -191,20 +208,18 @@ def invitation_partner(request, uuid):
                 request,
                 "Aquest enllaç d'invitació no existeix.",
             )
-            return redirect("edit_project")
+            return redirect("home")
         if request.user.id != invitation.user.id:
             messages.error(
                 request,
                 "Estàs accedint a un enllaç d'invitació a un projecte que correspon a un altre usuari. "
                 "Assegura't d'obrir l'enllaç des del compte a la qual t'han enviat la invitació.",
             )
-            return redirect("edit_project")
-        context = {
-            "project": invitation.project,
-            "user": invitation.user
-        }
+            return redirect("home")
+        context = {"project": invitation.project, "user": invitation.user}
         return render(request, "project_invitation.html", context=context)
     if request.method == "POST":
+        project_pk = request.POST.get("project")
         try:
             invitation = Invitation.objects.get(uuid=uuid)
         except Invitation.DoesNotExist:
@@ -212,7 +227,7 @@ def invitation_partner(request, uuid):
                 request,
                 "Aquest enllaç d'invitació no existeix.",
             )
-            return redirect("edit_project")
+            return redirect("home")
         try:
             user = User.objects.get(id=invitation.user.id)
         except User.DoesNotExist:
@@ -220,19 +235,31 @@ def invitation_partner(request, uuid):
                 request,
                 "Aquest usuari no existeix.",
             )
-            return redirect("edit_project")
+            return redirect("home")
         if "accept" in request.POST:
             try:
-                project = Project.objects.get(pk=request.POST.get("project"))
+                project = Project.objects.get(pk=project_pk)
             except Project.DoesNotExist:
                 messages.error(
                     request,
                     "Aquest projecte no existeix.",
                 )
-                return redirect("edit_project")
+                return redirect("home")
             project.partners.add(user)
-        invitation.delete()
-        return redirect("edit_project")
+            invitation.delete()
+            messages.error(
+                request,
+                "Has acceptat formar part del projecte.",
+            )
+            return redirect("edit_project", pk=project_pk)
+        if "deny" in request.POST:
+            invitation.delete()
+            messages.error(
+                request,
+                "Has rebutjat formar part del projecte.",
+            )
+            return redirect("home")
+    return redirect("home")
 
 
 class ProjectCreateFormView(SuccessMessageMixin, generic.CreateView):
@@ -242,18 +269,14 @@ class ProjectCreateFormView(SuccessMessageMixin, generic.CreateView):
     extra_context = {"show_new_project_info": True}
 
     def get_success_url(self):
-        return urls.reverse("edit_project")
+        return urls.reverse("edit_project", kwargs={"pk": self.object.pk})
 
     def form_valid(self, form):
         newproject = form.save()
         newproject.partners.add(self.request.user)
         messages.success(self.request, "Dades del projecte guardades correctament.")
+        self.object = newproject
         return HttpResponseRedirect(self.get_success_url())
-
-    def get(self, request):
-        if self.request.user.project is not None:
-            return HttpResponseRedirect(urls.reverse("edit_project"))
-        return super().get(self, request)
 
 
 class ProjectInfoView(LoginSignupContainerView):
